为email添加index和唯一性验证 (ch 6.2.5 Uniqueness validation) 添加index也可以提升检索性能，因为 based on the naïve data model, the only way to find a user by email address is to look through each user row in the database and compare its email attribute to the given email. This is known in the database business as a full-table scan, and for a real site with thousands of users it is a Bad Thing.
Putting an index on the email column fixes the problem. To understand a database index, it’s helpful to consider the analogy of a book index. In a book, to find all the occurrences of a given string, say “foobar”, you would have to scan each page for “foobar”. With a book index, on the other hand, you can just look up “foobar” in the index to see all the pages containing “foobar”. A database index works essentially the same way.